TypeScriptでhelloworldまで
