+++ pages/hello.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+const message: string = 'Hello, World!'
+</script>
+
+<template>
+  <div>
+    <h1>{{ message }}</h1>
+  </div>
+</template>
